tests/wpt/web-platform-tests/css/css-typed-om/stylevalue-subclasses/numeric-objects/cssMathNegate-type.html

<!doctype html>
<meta charset="utf-8">
<title>CSSMathNegate.type</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#type-of-a-cssmathvalue">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testhelper.js"></script>
<body>
<div id="log">
<script>
'use strict';

test(() => {
  const result = new CSSMathNegate(new CSSUnitValue(0, 'number'));
  assert_numeric_type_equals(result.type(), { });
}, 'Negating a type with empty map returns the empty map');

test(() => {
  const result = new CSSMathNegate(new CSSUnitValue(0, 'px'));
  assert_numeric_type_equals(result.type(), { length: 1 });
}, 'Negating a type returns the same type');

</script>
