LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/numeric-objects/cssMathNegate-type-expected.txt


FAIL Negating a type with empty map returns the empty map assert_equals: length expected (undefined) undefined but got (number) 0
FAIL Negating a type returns the same type assert_equals: length expected 1 but got 0
